Allow for using a local copy of the installer tarball
Prior to this commit, the installer required a download of the tarball.
As this is a limitation on some sites and puts more load on the
networking, this commit allow for using the tarball_installer option to
specify the location of the installer tarball on the target machine to
use.

This commit also changes the default base_url to use the externally
accessible download source.