Use the agent section of the puppet.conf to get certname
Prior to this commit, the main section of the puppet.conf was used to
get the certname for the servers. As puppet-bootstrap uses the agent
section to specify the certname, windows nodes were failing to provision
due to not being able to sign the certificate. This commit specifies
using the agent section instead of the main when getting the certname
used by the agent.