(fix) Hardcode the LANG for tasks
Prior to this commit, the LANG would be whatever is on the machine
running bolt. In general this works fine, however if there is a LANG
that the target machine does not have, the installation could fail. This
was seen when running a restart exec where there were non ASCII
characters in the systemd output. This commit hard codes the LANG to
en_US.UTF-8 in order to avoid issues.